Initial refactoring
