+++ qualibration_graphs/quantum_dots/quam_config/.qm_cluster_config.json
@@ -1,4 +1,4 @@
 {
-  "host": "172.16.33.114",
-  "cluster_name": "CS_4"
+    "host": "172.16.33.115",
+    "cluster_name": "CS_4"
 }
