--- qualibration_graphs/quantum_dots/validation_utils/charge_stability/__init__.py
@@ -1,8 +1,0 @@
-"""Charge stability validation utilities for quantum dots."""
-
-from .default import InitDotModel, init_dot_model
-
-__all__ = [
-    "InitDotModel",
-    "init_dot_model",
-]
