--- qualibration_graphs/quantum_dots/validation_utils/time_dynamics/__init__.py
@@ -1,8 +1,0 @@
-"""Time dynamics validation utilities for quantum dots."""
-
-from . import src
-from .src import *
-
-__all__ = [
-    *src.__all__,
-]
